Cloud Functionality

+++ app/src/main/java/com/dsk/myexpense/expense_module/data/repository/ExpenseRepository.kt
@@ -36,7 +36,7 @@ class ExpenseRepository(
     val allExpenseDetails: LiveData<List<ExpenseDetails>> = expenseDAO.getAllExpenseDetails()
     val getTotalIncomeAmount: LiveData<Double> = expenseDAO.getTotalIncome().asLiveData()
     val getTotalExpenseAmount: LiveData<Double> = expenseDAO.getTotalExpense().asLiveData()
-    val getTotalIncomeExpenseAmount: LiveData<Int> = expenseDAO.getTotalIncomeExpense().asLiveData()
+    val getTotalIncomeExpenseAmount: LiveData<Double> = expenseDAO.getTotalIncomeExpense().asLiveData()
 
     suspend fun saveExpenseWithInvoice(
         context: Context,
